Import of DDUMONT/Config-Model-TkUI-1.312 from CPAN.
gitpan-cpan-distribution: Config-Model-TkUI
gitpan-cpan-version:      1.312
gitpan-cpan-path:         DDUMONT/Config-Model-TkUI-1.312.tar.gz
gitpan-cpan-author:       DDUMONT
gitpan-cpan-maturity:     released

+++ lib/Config/Model/TkUI.pm
@@ -11,7 +11,7 @@
 
 package Config::Model::TkUI ;
 BEGIN {
-  $Config::Model::TkUI::VERSION = '1.311';
+  $Config::Model::TkUI::VERSION = '1.312';
 }
 
 use strict;
@@ -636,8 +636,10 @@ sub disp_hash {
 	# check for display order mismatch
 	if ($tkt->infoExists($newpath) and $prevpath) {
 	    my $tkprevpath = $tkt->info( prev => $newpath );
-	    $logger->trace("disp_hash deleting mismatching $newpath mode $eltmode cargo_type $elt_type" );
-	    $tkt->delete(entry => $newpath) ;
+	    if ($prevpath ne $tkprevpath) {
+		$logger->trace("disp_hash deleting mismatching $newpath mode $eltmode cargo_type $elt_type" );
+		$tkt->delete(entry => $newpath) ;
+	    }
 	}
 
 	# check for content mismatch
